refactor: swipe features로 분리 (#51)
