+++ src/features/swipe/index.ts
@@ -1,0 +1,1 @@
+export { default as SwipeCarousel } from './ui/SwipeCarousel'
